move UnsupervisedTrainingMixin to init
