+++ scvi/external/solo/_model.py
@@ -15,7 +15,7 @@ from scvi.data import get_from_registry, setup_anndata
 from scvi.dataloaders import AnnDataLoader
 from scvi.lightning import ClassifierTrainingPlan
 from scvi.model import SCVI
-from scvi.model.base import BaseModelClass
+from scvi.model.base import BaseModelClass, UnsupervisedTrainingMixin
 from scvi.modules import Classifier
 
 logger = logging.getLogger(__name__)
@@ -23,7 +23,7 @@ logger = logging.getLogger(__name__)
 LABELS_KEY = "_solo_doub_sim"
 
 
-class SOLO(BaseModelClass):
+class SOLO(UnsupervisedTrainingMixin, BaseModelClass):
     """
     Doublet detection in scRNA-seq [Bernstein19]_.
