+++ scvi/model/_autozi.py
@@ -11,7 +11,7 @@ from scvi import _CONSTANTS
 from scvi._compat import Literal
 from scvi.dataloaders import AnnDataLoader
 from scvi.lightning import TrainingPlan
-from scvi.model.base._training_mixin import UnsupervisedTrainingMixin
+from scvi.model.base import UnsupervisedTrainingMixin
 from scvi.modules import AutoZIVAE
 
 from .base import BaseModelClass, VAEMixin
