+++ scvi/model/_peakvi.py
@@ -19,7 +19,7 @@ from scvi.model._utils import (
     _get_var_names_from_setup_anndata,
     scatac_raw_counts_properties,
 )
-from scvi.model.base._training_mixin import UnsupervisedTrainingMixin
+from scvi.model.base import UnsupervisedTrainingMixin
 from scvi.modules import PEAKVAE
 
 from .base import ArchesMixin, BaseModelClass, VAEMixin
