+++ scvi/model/base/__init__.py
@@ -1,6 +1,7 @@
 from ._archesmixin import ArchesMixin
 from ._base_model import BaseModelClass
 from ._rnamixin import RNASeqMixin
+from ._training_mixin import UnsupervisedTrainingMixin
 from ._vaemixin import VAEMixin
 
 __all__ = [
@@ -8,4 +9,5 @@ __all__ = [
     "BaseModelClass",
     "RNASeqMixin",
     "VAEMixin",
+    "UnsupervisedTrainingMixin",
 ]
